Open alert messages when deleting or exiting from main page

+++ src/inventory/gui/MainScreenController.java
@@ -8,12 +8,14 @@ package inventory.gui;
 import inventory.models.Part;
 import inventory.models.Inventory;
 import static inventory.models.Inventory.deletePart;
+import static inventory.models.Inventory.deleteProduct;
 import static inventory.models.Inventory.getAllParts;
 import static inventory.models.Inventory.getAllProducts;
 import inventory.models.Product;
 import java.io.IOException;
 import java.net.URL;
 import java.text.NumberFormat;
+import java.util.Optional;
 import java.util.ResourceBundle;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
@@ -23,11 +25,14 @@ import javafx.fxml.Initializable;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
+import javafx.stage.Modality;
 import javafx.stage.Stage;
 
 /**
@@ -89,9 +94,34 @@ public class MainScreenController implements Initializable {
     
     @FXML
     private void deletePartFromInventory(ActionEvent event) {
-        Part selectedPart = partsTable.getSelectionModel().getSelectedItem();
-        
-        deletePart(selectedPart);
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.initModality(Modality.NONE);
+        alert.setTitle("Part Delete Confirmation");
+        alert.setHeaderText("Confirm deletion of part");
+        alert.setContentText("Are you sure you want to delete part?");
+        Optional<ButtonType> result = alert.showAndWait();
+
+        if (result.get() == ButtonType.OK) {
+            Part selectedPart = partsTable.getSelectionModel().getSelectedItem();
+            deletePart(selectedPart);
+            updatePartsTable();
+        }    
+    }
+    
+    @FXML
+    private void deleteProductFromInventory() {
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.initModality(Modality.NONE);
+        alert.setTitle("Product Delete Confirmation");
+        alert.setHeaderText("Confirm deletion of product");
+        alert.setContentText("Are you sure you want to delete part?");
+        Optional<ButtonType> result = alert.showAndWait();
+
+        if (result.get() == ButtonType.OK) {
+            Product selectedProduct = productsTable.getSelectionModel().getSelectedItem();
+            deleteProduct(selectedProduct);
+            updateProductsTable();
+        }   
     }
     
     @FXML
@@ -147,8 +177,15 @@ public class MainScreenController implements Initializable {
     
     @FXML
     private void closeMain(ActionEvent event) {
-        Stage stage = (Stage) closeMainBtn.getScene().getWindow();
-        stage.close();
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.initModality(Modality.NONE);
+        alert.setTitle("Confirm Exit");
+        alert.setHeaderText("Please confirm you want to exit.");
+        alert.setContentText("Any unsaved changes will be lost.");
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.get() == ButtonType.OK) {
+            System.exit(0);
+        }
     }
     
     @Override
